Replace files indices.h/cpp with index_vector.h/cpp, index_vector_full.h/cpp and index_vector_partial.h/cpp.
