+++ python/boomer/boosting/cpp/data/matrix_dense_example_wise.h
@@ -3,7 +3,7 @@
  */
 #pragma once
 
-#include "../../../common/cpp/indices.h"
+#include "../../../common/cpp/data/types.h"
 
 
 namespace boosting {
