+++ python/boomer/common/cpp/indices/index_vector_partial.h
@@ -1,0 +1,78 @@
+/**
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "index_vector.h"
+#include "../data/vector_dense.h"
+
+
+/**
+ * Provides random access to a fixed number of indices stored in a C-contiguous array.
+ */
+class PartialIndexVector : public IIndexVector {
+
+    private:
+
+        DenseVector<uint32> vector_;
+
+    public:
+
+        /**
+         * @param numElements The number of elements in the vector
+         */
+        PartialIndexVector(uint32 numElements);
+
+        typedef DenseVector<uint32>::iterator iterator;
+
+        typedef DenseVector<uint32>::const_iterator const_iterator;
+
+        /**
+         * Returns an `iterator` to the beginning of the indices.
+         *
+         * @return An `iterator` to the beginning
+         */
+        iterator begin();
+
+        /**
+         * Returns an `iterator` to the end of the indices.
+         *
+         * @return An `iterator` to the end
+         */
+        iterator end();
+
+        /**
+         * Returns a `const_iterator` to the beginning of the indices.
+         *
+         * @return A `const_iterator` to the beginning
+         */
+        const_iterator cbegin() const;
+
+        /**
+         * Returns a `const_iterator` to the end of the indices.
+         *
+         * @return A `const_iterator` to the end
+         */
+        const_iterator cend() const;
+
+        /**
+         * Sets the number of indices.
+         *
+         * @param numElements The number of indices to be set
+         */
+        void setNumElements(uint32 numElements);
+
+        uint32 getNumElements() const override;
+
+        bool isPartial() const override;
+
+        uint32 getIndex(uint32 pos) const override;
+
+        std::unique_ptr<IStatisticsSubset> createSubset(const IHistogram& histogram) const override;
+
+        std::unique_ptr<IRuleRefinement> createRuleRefinement(IThresholdsSubset& thresholdsSubset,
+                                                              uint32 featureIndex) const override;
+
+        std::unique_ptr<IHeadRefinement> createHeadRefinement(const IHeadRefinementFactory& factory) const override;
+
+};
