+++ python/boomer/seco/cpp/rule_evaluation_label_wise.h
@@ -7,8 +7,9 @@
  */
 #pragma once
 
+#include "../../common/cpp/indices/index_vector_full.h"
+#include "../../common/cpp/indices/index_vector_partial.h"
 #include "../../common/cpp/rule_evaluation.h"
-#include "../../common/cpp/indices.h"
 #include "heuristics.h"
 #include <memory>
