+++ python/setup.py
@@ -19,8 +19,9 @@ sources = [
     'boomer/common/cpp/data/vector_dense.cpp',
     'boomer/common/cpp/data/vector_dok_binary.cpp',
     'boomer/common/cpp/data/vector_sparse_array.cpp',
+    'boomer/common/cpp/indices/index_vector_full.cpp',
+    'boomer/common/cpp/indices/index_vector_partial.cpp',
     'boomer/common/cpp/random.cpp',
-    'boomer/common/cpp/indices.cpp',
     'boomer/common/cpp/input_data.cpp',
     'boomer/common/cpp/predictions.cpp',
     'boomer/common/cpp/rules.cpp',
